Add method to get the shift applied by the homography for an input point
Also encapsulates the functionality of adapting the input point (so it
can be either a Shapely point or a 1D 2 element numpy array).

Added test for the new method.

Rest of changes include code quality issues (autopep8 applied) and pt
renamed to point in the whole file to have coherency.